Add Review Data 초기 리뷰 데이터 적재 및 product id 업데이트

+++ camping/src/main/java/com/camping/dev/model/entity/Review.java
@@ -15,7 +15,7 @@ public class Review {
     private String email;
 
     // 리뷰 평점
-    private Long grade;
+    private int grade;
 
     // 긍부정 판별
     private Boolean result_yn;
